fix(backend): ai 预测/工艺优化表 created_by 等列模型 i64→i32 对齐实库 INT4 修 DATABASE_ERROR
- POST /ai/quality-predictions 报 `decoding column created_by: Option<i64> 不兼容 SQL INT4`
  根因: users/products.id 权威是 i32/SERIAL、仓库 80+ 处 created_by 皆 Option<i32>，唯两 ai 表模型误用 i64；
  m0044 fix_fk_types() 运行期把 BIGINT REFERENCES 静默改成 INTEGER，故实库=INT4，模型却 i64
- 方向: 改模型对齐实库/约定(不 ALTER 实列，否则与 users.id INT4 的 FK 冲突): ai_quality_prediction(product_id/acknowledged_by/created_by)、ai_process_optimization(applied_by/created_by) i64→i32
- 迁移文本 5 处 BIGINT→INTEGER 自洽化(运行期行为不变);service 去 .map(as i64)/.as i32 冗余、handler 去 auth.user_id as i64
- cargo fmt --check 0;只动 backend/。遗留同类漂移 product_color_price.rs:30 created_by:i64 下一批同向修

+++ backend/src/models/ai_quality_prediction.rs
@@ -11,7 +11,8 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
     pub request_id: String,
-    pub product_id: Option<i64>,
+    /// 关联 products.id（INTEGER/SERIAL 主键）
+    pub product_id: Option<i32>,
     pub inspection_type: String,
     pub window_days: i32,
     pub total_inspections: i64,
@@ -27,7 +28,8 @@ pub struct Model {
     pub source: String,
     pub is_acknowledged: bool,
     pub acknowledged_at: Option<DateTime<Utc>>,
-    pub acknowledged_by: Option<i64>,
+    /// 关联 users.id（INTEGER/SERIAL 主键）
+    pub acknowledged_by: Option<i32>,
     /// V15 P1 2.1：实际风险等级（对账回填）
     pub actual_risk_level: Option<String>,
     /// V15 P1 2.1：实际平均合格率（对账回填）
@@ -44,7 +46,8 @@ pub struct Model {
     pub claim_amount: Option<Decimal>,
     /// V15 P2 14.2.3：索赔记录时间
     pub claim_recorded_at: Option<DateTime<Utc>>,
-    pub created_by: Option<i64>,
+    /// 关联 users.id（INTEGER/SERIAL 主键）
+    pub created_by: Option<i32>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
